Move nunjucks services to a directory
